Update 008 - Conversor de Medidas.py

+++ 008 - Conversor de Medidas.py
@@ -1,4 +1,4 @@
-m =int(input('Distância em metros: '))
+m = float(input('Distância em metros: '))
 cm = m * 100
 ml = m * 1000
 print('A distância de {}m, contém {}cm e {}ml em medidas!'.format(m, cm, ml))
